feat(player): 플레이어 구현
- 플레이어 파워업 및 공격 속도에 따라 다른 공격 구현
- 최적화를 위한 플레이어 총알 오브젝트 풀링 구현
- 플레이어 움직임 추가
- 플레이어 데미지 구현
- 플레이어 사망 시 폭발 애니메이션 및 움직임 구현

+++ AGK.ShootingGame/Assets/Scripts/Manager/GameManager.cs
@@ -29,14 +29,14 @@ public class GameManager : MonoBehaviour
 
     public void StartGame()
     {
-        SceneManager.LoadScene(Scenes.Stage1);
+        SceneManager.LoadScene((int)Scenes.Stage1);
     }
 
     public void MoveToNextScene()
     {
-        if (currentScene == Scenes.LAST_STAGE)
+        if (currentScene == (int)Scenes.LAST_STAGE)
         {
-            SceneManager.LoadScene(Scenes.End);
+            SceneManager.LoadScene((int)Scenes.End);
             return;
         }
         
@@ -45,6 +45,6 @@ public class GameManager : MonoBehaviour
 
     public void GameOver()
     {
-        SceneManager.LoadScene(Scenes.GameOver);
+        SceneManager.LoadScene((int)Scenes.GameOver);
     }   
 }
